Expanded Industry
Some reorganization of main mod files to put all similar actions near
each other (crafting recipes, smelting, etc)
Added harvest craft integration
Updated some tooltips to include mod ID
Additional achievement unlock conditions
Expanded Industry now working + assets for all other blocks to come

+++ src/main/java/com/draco18s/farming/item/ItemHydrometer.java
@@ -83,8 +83,8 @@ public class ItemHydrometer extends Item {
 			tag.setBoolean("HasOffsets", true);
 			tag.setFloat("rainflat", offsets.rainfallFlat);
 			tag.setFloat("tempflat", offsets.temperatureFlat);
-			tag.setInteger("raintime", offsets.rainfallTimeOffset);
-			tag.setInteger("temptime", offsets.temperatureTimeOffset);
+			tag.setFloat("raintime", offsets.rainfallTimeOffset);
+			tag.setFloat("temptime", offsets.temperatureTimeOffset);
 			
 			Item item = Item.getItemFromBlock(block);
 			if (block == Blocks.WHEAT){
